Resolve conflicts after merge with main
